requirements.txt
# Core
python-dotenv>=1.0

# ML
scikit-learn>=1.3
xgboost>=2.0
imbalanced-learn>=0.11
optuna>=3.4

# Explainability
shap>=0.43

# API
fastapi>=0.104
uvicorn>=0.24
pydantic>=2.5

# Dashboard
streamlit>=1.28
plotly>=5.18

# Data
pandas>=2.1
numpy>=1.26
pyarrow>=14.0

# Database
aiosqlite>=0.19

# Utilities
pyyaml>=6.0
joblib>=1.3

# Dev
pytest>=7.0
pytest-cov>=4.0
pytest-asyncio>=0.23
ruff>=0.1
pre-commit>=3.0
